Truncate output build info file
In certain cases, the output file can "shrink" compared to the input
file (for example, when download URLs are verified and their status
changes to `true` from `false` (length reduction). This truncate step
ensures that the output file is right-sized and has no leftover text
from before, when the output file is the same as the input file.